fix: RunSpace was unusable — scrim over it, terminal leaking, dead menu
Five defects reported together. Four share one cause and one is separate.

── 1. The sheet over RunSpace, and the dead menu button ──────────────────
.rs-menu-backdrop shipped as

    position: fixed; inset: 0; z-index: 855; background: rgba(0,0,0,.6);

with nothing anywhere that hides it. #wbBackdrop is the FIRST child of
#tab-jobs, so opening RunSpace dropped an opaque sheet across the whole
view. That is also the whole of "মেনু বাটনে ক্লিক করলে কোন কাজ হয় না": the
button works, but the scrim is above it at z-index 855 and eats the click,
and the rail the button opens was underneath the scrim anyway. Nothing in
the view could be reached.

It now starts visibility:hidden / pointer-events:none, and appears only
under 760px with body.rs-side-open. Desktop gets no scrim at all: there
the rail is a column beside the editor, not something laid over it.

── 2. The terminal and its keypad on top of the editor ───────────────────
    .dash-tab-content { display: none }        (line 442)
    .term, .term-page, .term-standalone { display: flex }   (line 1105)

The second came later and won. #tab-term is a .dash-tab-content, so the
standalone terminal page rendered inside whichever tab was open — its
ESC/TAB/CTRL keypad included. That is the "অগোছালো লেখা" under the editor
and the stray tab/ESC controls. The terminal page now obeys the tab rules
and the keypad is scoped to it.

── 3. Desktop and mobile fighting ────────────────────────────────────────
.rs-side is a 268px column at :root and a fixed off-canvas drawer under
760px, but the collapse rule existed only in the desktop form (width:0).
pro.js sets rs-side-open AND rs-side-collapsed together on <body>, derived
from the measured width — so on a phone the collapsed class landed with no
rule to answer it and the drawer looked stuck. Both widths now define both
states, and width:0 is not used on mobile because a drawer slides rather
than shrinks.

── 4. New shell: "uncaught parameter type" ───────────────────────────────
Reproduced, not guessed. POST /api/terminals with the body the frontend
actually sends returns 422:

    {"type": "int_from_float",
     "msg": "Input should be a valid integer, got a number with a
             fractional part"}

cols/rows come from xterm's FitAddon, which divides the pane box by the
measured cell size WITHOUT rounding: a 100px pane with a 9.6px cell is
10.416 columns. It is fractional whenever the viewport is not an exact
multiple of the cell — most often on a phone, where the browser toolbar
and the on-screen keyboard resize the pane constantly, which is why it
looked intermittent. terminal.js then retries the same rejected body every
2.5s forever, so the shell never opens.

TermCreateRequest.cols/rows are float now and the handler floors after
clamping. Same fix applied to runner/app.py's TerminalCreateRequest, which
had the identical annotation. The WebSocket resize path already wrapped
int() and needed nothing. Verified across normal / fractional / very
fractional / null / string inputs — all five now resolve to whole cells.

── Proof the guard is real ───────────────────────────────────────────────
tests/js/test_runspace_not_blocked.js builds the page at 1280px and 720px
by lifting the matching @media blocks (jsdom ignores them) and drives the
rail through open -> collapsed -> open at both widths. With the four CSS
fixes reverted it reports the reported symptoms verbatim:

    FAIL it does NOT block the view by default -> pos=fixed vis=visible pe=auto
    FAIL it does not swallow clicks
    FAIL terminal page is hidden while another tab is active -> flex
    FAIL the ESC/TAB keypad is hidden with it

and 32/32 with them in place.

Verified: 33/33 JS suites, test_all_routes 65, test_admin_dashboard 145,
test_system_tools 46.

+++ services/term_proxy.py
@@ -26,8 +26,26 @@ log = logging.getLogger("term_proxy")
 
 class TermCreateRequest(BaseModel):
     shell: Optional[str] = "bash"
-    cols: Optional[int] = 90
-    rows: Optional[int] = 28
+    # cols/rows are floats on the wire, not ints.
+    #
+    # REPRODUCED: creating a shell failed with a 422 whose body is
+    #   {"type": "int_from_float",
+    #    "msg": "Input should be a valid integer, got a number with a
+    #            fractional part"}
+    # which the frontend surfaces as an uncaught parameter-type error and
+    # then silently retries forever, so the terminal never opens.
+    #
+    # The value comes from xterm's FitAddon, which divides the element box
+    # by the measured cell size and does NOT round: a 100px-wide pane with
+    # a 9.6px cell is 10.416 columns. It lands fractional whenever the
+    # viewport is not an exact multiple of the cell -- most often on a
+    # phone, where the browser toolbar and the on-screen keyboard resize
+    # the pane constantly. That is why it looked intermittent.
+    #
+    # float accepts both 80 and 80.5; the handler floors it. Rejecting the
+    # request over a rendering detail the user cannot control is the bug.
+    cols: Optional[float] = 90
+    rows: Optional[float] = 28
     slot: Optional[int] = None
     name: Optional[str] = ""
     persist: Optional[bool] = None
@@ -53,8 +71,10 @@ def create_terminal(payload: TermCreateRequest, authorization: Optional[str] = H
     shell = (payload.shell or "bash").strip().lower()
     if shell not in ("bash", "python", "node"):
         raise HTTPException(400, "shell must be bash / python / node")
-    cols = max(40, min(payload.cols or 90, 240))
-    rows = max(10, min(payload.rows or 28, 80))
+    # Floor after clamping: a PTY takes whole cells, and int() on a value
+    # that is already inside the range cannot push it back out.
+    cols = int(max(40, min(payload.cols or 90, 240)))
+    rows = int(max(10, min(payload.rows or 28, 80)))
     slot = payload.slot if payload.slot and payload.slot > 0 else None
 
     try:
